feat(SDK): ability to synchronise the play area transform on sdk switch
A new checkbox on the SDK Setup Switcher that when checked will
synchronise the play area transform to the previous play area transform
when the SDK is switched.

This is useful when wanting to move the play area around with the
Simulator and then switch to an SDK of an actual HMD and be in the
position of where the Simulator was moved to.